+++ web/modules/custom/hoeringsportal_citizen_proposal/src/Exception/CprException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Drupal\hoeringsportal_citizen_proposal\Exception;
+
+/**
+ * CPR exception.
+ */
+class CprException extends RuntimeException {
+}